BioStructureViewer: Move code parts into molstar-highlight-utils.ts
